% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{edges2sg}
\alias{edges2sg}
\alias{nodes2sg}
\title{Prepare}
\usage{
edges2sg(x)

nodes2sg(x)
}
\arguments{
\item{x}{A data.frame of edges or nodes.}
}
\description{
Prepare \code{graphTweets} data for \code{sigmajs}.
}
